Update mailbox.c
integrated the fix proposed in pimoroni/unicorn-hat@5c03f81) to get the library to work with newer kernels.

+++ mailbox.c
@@ -11,7 +11,7 @@ modification, are permitted provided that the following conditions are met:
       documentation and/or other materials provided with the distribution.
     * Neither the name of the copyright holder nor the
       names of its contributors may be used to endorse or promote products
-      derived from this software without specific prior written permission.
+      derived from this software without specific pråior written permission.
 
 THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
 ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
@@ -270,19 +270,36 @@ int mbox_open(void) {
    char filename[64];
 
    // open a char device file used for communicating with kernel mbox driver
-   sprintf(filename, "/tmp/mailbox-%d", getpid());
-   unlink(filename);
-   if (mknod(filename, S_IFCHR|0600, makedev(100, 0)) < 0) {
-      printf("Failed to create mailbox device %s: %m\n", filename);
-      return -1;
-   }
+   //sprintf(filename, "/tmp/mailbox-%d", getpid());
+   //unlink(filename);
+   //if (mknod(filename, S_IFCHR|0600, makedev(100, 0)) < 0) {
+     // printf("Failed to create mailbox device %s: %m\n", filename);
+     // return -1;
+   //}
+   
+	
+   sprintf(filename, "/dev/vcio");
+	
    file_desc = open(filename, 0);
+	
+	if( file_desc < 0 ){
+		printf("Failed to open %s, trying old method.\n", filename);
+		sprintf(filename, "/dev/mailbox-%d", getpid());
+		unlink(filename);
+		if (mknod(filename, S_IFCHR|0600, makedev(100, 0)) < 0) {
+			printf("Failed to create mailbox device %s: %m\n", filename);
+			return -1;
+		}
+		file_desc = open(filename, 0);
+	}
+	
+	
    if (file_desc < 0) {
       printf("Can't open device file %s: %m\n", filename);
-      unlink(filename);
+      //unlink(filename);
       return -1;
    }
-   unlink(filename);
+   //unlink(filename);
 
    return file_desc;
 }
